+++ nomos-da/kzgrs-backend-testutils/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "kzgrs-backend-testutils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+kzgrs-backend = { path = "../kzgrs-backend" }
+nomos-core = { path = "../../nomos-core/chain-defs" }
