+++ nomos-da/network/core/Cargo.toml
@@ -28,4 +28,5 @@ thiserror = "1.0"
 tokio = { version = "1", features = ["macros", "rt-multi-thread", "time"] }
 libp2p = { version = "0.53", features = ["ed25519", "ping", "macros", "quic", "tcp", "yamux", "noise"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+kzgrs-backend-testutils = { path = "../../kzgrs-backend-testutils" }
