+++ nomos-da/network/messages/Cargo.toml
@@ -12,3 +12,4 @@ tokio = "1"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["macros"] }
+kzgrs-backend-testutils = { path = "../../kzgrs-backend-testutils" }
