+++ nomos-da/network/messages/src/packing.rs
@@ -104,50 +104,13 @@ mod tests {
     use crate::common::Blob;
     use crate::dispersal::{DispersalError, DispersalErrorType, DispersalRequest};
     use futures::io::BufReader;
-    use kzgrs_backend::common::blob::DaBlob;
-    use kzgrs_backend::encoder::{self, DaEncoderParams};
-    use nomos_core::da::{BlobId, DaEncoder};
-
-    fn get_encoder() -> encoder::DaEncoder {
-        const DOMAIN_SIZE: usize = 16;
-        let params = DaEncoderParams::default_with(DOMAIN_SIZE);
-        encoder::DaEncoder::new(params)
-    }
-
-    fn get_da_blob() -> DaBlob {
-        let encoder = get_encoder();
-        let data = vec![
-            49u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8,
-            0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8, 0u8,
-        ];
-
-        let encoded_data = encoder.encode(&data).unwrap();
-        let columns: Vec<_> = encoded_data.extended_data.columns().collect();
-
-        let index = 0;
-        let da_blob = DaBlob {
-            column: columns[index].clone(),
-            column_idx: index
-                .try_into()
-                .expect("Column index shouldn't overflow the target type"),
-            column_commitment: encoded_data.column_commitments[index],
-            aggregated_column_commitment: encoded_data.aggregated_column_commitment,
-            aggregated_column_proof: encoded_data.aggregated_column_proofs[index],
-            rows_commitments: encoded_data.row_commitments.clone(),
-            rows_proofs: encoded_data
-                .rows_proofs
-                .iter()
-                .map(|proofs| proofs.get(index).cloned().unwrap())
-                .collect(),
-        };
-
-        da_blob
-    }
+    use kzgrs_backend_testutils::get_da_blob;
+    use nomos_core::da::BlobId;
 
     #[tokio::test]
     async fn pack_and_unpack() -> Result<()> {
         let blob_id = BlobId::from([0; 32]);
-        let data = get_da_blob();
+        let data = get_da_blob(None);
         let blob = Blob::new(blob_id, data);
         let subnetwork_id = 0;
         let message = DispersalRequest::new(blob, subnetwork_id);
@@ -162,7 +125,7 @@ mod tests {
     #[tokio::test]
     async fn pack_to_writer_and_unpack_from_reader() -> Result<()> {
         let blob_id = BlobId::from([0; 32]);
-        let data = get_da_blob();
+        let data = get_da_blob(None);
         let blob = Blob::new(blob_id, data);
         let subnetwork_id = 0;
         let message = DispersalRequest::new(blob, subnetwork_id);
